# apps/web/ui/lib/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ComponentStatus": {
        "additionalProperties": false,
        "description": "模块声明的一个依赖组件的就绪状态（如实探测，不伪造 ready）。",
        "properties": {
          "available": {
            "default": false,
            "title": "Available",
            "type": "boolean"
          },
          "detail": {
            "default": "",
            "title": "Detail",
            "type": "string"
          },
          "module_path": {
            "default": "",
            "title": "Module Path",
            "type": "string"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          },
          "required": {
            "default": true,
            "title": "Required",
            "type": "boolean"
          },
          "schema_version": {
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "title": "ComponentStatus",
        "type": "object"
      },
...
      "DomainModuleDescriptor": {
        "additionalProperties": false,
        "description": "领域模块声明（任务书 6.8 字段清单）。",
        "properties": {
          "api_routes": {
            "items": {
              "type": "string"
            },
            "title": "Api Routes",
            "type": "array"
          },
          "artifact_types": {
            "items": {
              "type": "string"
            },
            "title": "Artifact Types",
            "type": "array"
          },
          "command_handlers": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Command Handlers",
            "type": "object"
          },
          "default_executor": {
            "default": "",
            "title": "Default Executor",
            "type": "string"
          },
          "event_namespaces": {
            "items": {
              "type": "string"
            },
            "title": "Event Namespaces",
            "type": "array"
          },
          "gate_binding": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gate Binding"
          },
          "graph_binding": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graph Binding"
          },
          "id": {
            "default": "",
            "title": "Id",
            "type": "string"
          },
          "required_capabilities": {
            "items": {
              "type": "string"
            },
            "title": "Required Capabilities",
            "type": "array"
          },
          "schema_version": {
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "task_kinds": {
            "items": {
              "type": "string"
            },
            "title": "Task Kinds",
            "type": "array"
          },
          "ui_contributions": {
            "additionalProperties": true,
            "title": "Ui Contributions",
            "type": "object"
          },
          "version": {
            "default": "",
            "title": "Version",
            "type": "string"
          },
          "workspace_kind": {
            "default": "",
            "title": "Workspace Kind",
            "type": "string"
          }
        },
        "title": "DomainModuleDescriptor",
        "type": "object"
      },
...
      "ErrorCategory": {
        "description": "错误分类，调用方据此决定重试、升级或放弃。",
        "enum": [
          "validation",
          "permission",
          "not_found",
          "conflict",
          "state",
          "rate_limit",
          "timeout",
          "runtime",
          "platform",
          "internal"
        ],
        "title": "ErrorCategory",
        "type": "string"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "统一错误 envelope。",
        "properties": {
          "category": {
            "$ref": "#/components/schemas/ErrorCategory",
            "default": "internal"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "detail": {
            "additionalProperties": true,
            "title": "Detail",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "recovery_hint": {
            "default": "",
            "title": "Recovery Hint",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          },
          "schema_version": {
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModuleRegistration": {
        "additionalProperties": false,
        "description": "一次模块注册的结果记录（含失败记录）。",
        "properties": {
          "components": {
            "items": {
              "$ref": "#/components/schemas/ComponentStatus"
            },
            "title": "Components",
            "type": "array"
          },
          "descriptor": {
            "$ref": "#/components/schemas/DomainModuleDescriptor"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorEnvelope"
              },
              {
                "type": "null"
              }
            ]
          },
          "origin": {
            "default": "builtin",
            "title": "Origin",
            "type": "string"
          },
          "pending_components": {
            "items": {
              "type": "string"
            },
            "title": "Pending Components",
            "type": "array"
          },
          "schema_version": {
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "state": {
            "$ref": "#/components/schemas/ModuleState",
            "default": "registered"
          }
        },
        "title": "ModuleRegistration",
        "type": "object"
      },
      "ModuleState": {
        "description": "模块生命周期状态。\n\n- ``registered``：描述校验通过，但声明的依赖组件未全部就绪。\n- ``ready``：校验通过且依赖齐全，可以提供 workspace kind 入口。\n- ``unavailable``：注册校验失败，``error`` 中记录明确原因。\n- ``disabled``：Operator 禁用；入口隐藏，记录与历史数据仍可读。",
        "enum": [
          "registered",
          "ready",
          "unavailable",
          "disabled"
        ],
        "title": "ModuleState",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Project Muteki — Agent 安全工作台",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/domain-modules/{module_id}": {
      "get": {
        "operationId": "get_domain_module_api_domain_modules__module_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "module_id",
            "required": true,
            "schema": {
              "title": "Module Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleRegistration"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Domain Module",
        "tags": [
          "platform-registry"
        ]
      }
    },
    "/api/domain-modules/{module_id}/{action}": {
      "post": {
        "operationId": "change_domain_module_api_domain_modules__module_id___action__post",
        "parameters": [
          {
            "in": "path",
            "name": "module_id",
            "required": true,
            "schema": {
              "title": "Module Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "action",
            "required": true,
            "schema": {
              "title": "Action",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "default": {},
                "title": "Body",
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change Domain Module",
        "tags": [
          "platform-registry"
        ]
      }
    },
...
  }
}